Project Update 1!
6:58 to 8:12pm; 74 min

I added my first two problem solutions, the first one finds the sum of all multiples of 3 or 5 and the second one finds the sum of all even fibonacci numbers less than 4 million. 

Some roadblocks I ran into were that    not in   didn't work for lists, so I finally turned the list into a set so that all duplicate terms would be removed. I also at first had 
   for i in range(333)   which didn't give me 999 and made me wonder why my sum wasn't right.

For the second function it added fibonacci values greater than 4 million to my list (which made sense since I was checking if the last index was less than 4 million, not the sum of that index and the one before it), so I solved this by simply only finding the sum of values in my list that were less than 4 million. I also misread the question at first and thought it said less than 3 million for some reason so I was confused why my answer wasn't right.