New Update!
I worked on this for a little over 75 minutes, (from about 9:40am to 11:00am)
I added SmallestMult(), SumSquareDiff(), BigPrime(), and BigProduct(), plus the helper function IsPrime() to help BigPrime(). SmallestMult(), SumSquareDiff(), and BigPrime() were relatively straightforward, with minimal roadblocks, besides that I got confused with the "and" vs "or" for a bit, and was wondering why SmallestMult() kept returning 2520 as the answer. BigProduct(), was a bit more challenging, since keeping the windows straight (and making sure the indices weren't off by one, etc.) was a bit confusing. I also misread the question twice, or misread that I misread it, since I was on the right track the first time and thought that it was different for some reason.