Update!
First of all, I realized that last time I forgot to add comments to some of the functions, so I replaced them with the commented versions. 
This last work session I programmed for about 50 minutes (from about 10:30 am to 11:20 am) and added PythTriplets and SumPrimes. A MASSIVE roadblock I faced was that I couldn't figure out another way to find all primes up to 2 million, so I did it the brute force way, and it took approximately 8 or so hours to run and just finished about 20 minutes ago. Thank goodness the answer was correct the first time, if it didn't work I was going to make a list of all numbers up to 2 million and remove non-primes from it. I'd still really like to know what the intended method was supposed to be.