New Update!
I spent 60 minutes total (15 min last night from about 9:15pm - 9:30pm, plus about 30 min from about 8:00am - 8:30am, plus an additional 15 minutes from about 10:30am - 10:45am).
I added the function GridProduct, which takes in a grid of numbers and finds the largest product on any diagonal (since I knew that they wouldn't make the problem so easy as to put the answer on one of the rows or columns haha). This function required a lot of prep (it took a while to get the huge list of numbers as lists for each of the 20 rows, and then I made GridList a list of the lists). It was really satisfying once I got the list of lists though, since I could find the number on any coordinate point. From here I did some nested loops and began getting indices of [i][j],[i+1][j+1] etc. to try line by line of a smaller grid I'd created (to avoid the diagonals with less than 4 number). A small roadblock I faced was that the coordinate system increases up to down, so it messed me up a few times and made it more confusing. The second diagonal (right to left) was also confusing to program since then I had to skip a couple i's right at the beginning to not get the diagonals less than 4.