Final Update!
I spent about 85 minutes from about 11:15 to 12:40 and added the function TriangleNums. I faced a few different roadblocks on this one, since keeping the different variables straight was challenging (such as figuring out that it should be TriangleNum = TriangleNum + i to generate TriangleNums). I also forgot to reset the count at the beginning of the loop at first so I was wondering why it kept increasing so much. Lastly, the big roadblock I faced was that I had no idea that it would be that many TriangleNums, and it's still running now searching through them. I'm sure there's an easier way, but I'm not sure what else I could do instead of counting the number of divisors of the bottom half of all of the TriangleNums. I'm pretty sure my program will come up with the right answer but it's still running.